requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
datasets==3.0.1
dateparser==1.2.0
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.5
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
GoogleNews==1.6.15
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==3.0.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
openai==1.51.2
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
pyperclip==1.9.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
selenium==4.25.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
sympy==1.13.3
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.15.1
